Moved test services -> TestServices project
