+++ src/Microsoft.VisualStudio.ProjectSystem.Managed.TestServices.UnitTests/Microsoft.VisualStudio.ProjectSystem.Managed.TestServices.UnitTests.csproj
@@ -44,5 +44,11 @@
   <ItemGroup>
     <None Include="project.json" />
   </ItemGroup>
+  <ItemGroup>
+    <Compile Include="ProjectSystem\ProjectTreeParserTests.cs" />
+  </ItemGroup>
+  <ItemGroup>
+    <Service Include="{82A7F48D-3B50-4B1E-B82E-3ADA8210C358}" />
+  </ItemGroup>
   <Import Project="..\..\build\Targets\VSL.Imports.targets" />
 </Project>
