+++ src/Microsoft.VisualStudio.ProjectSystem.Managed.TestServices/ProjectSystem/ProjectSystemTraitDiscover.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using Xunit.Abstractions;
 using Xunit.Sdk;
 
-namespace Microsoft.VisualStudio.Testing
+namespace Microsoft.VisualStudio
 {
     public class ProjectSystemTraitDiscoverer : ITraitDiscoverer
     {
