+++ src/Microsoft.VisualStudio.ProjectSystem.Managed.TestServices/ProjectSystem/ProjectTreeParser/ProjectTreeParser.cs
@@ -3,9 +3,9 @@
 using System;
 using System.Collections.Immutable;
 using Microsoft.VisualStudio.ProjectSystem;
-using static Microsoft.VisualStudio.Testing.Tokenizer;
+using static Microsoft.VisualStudio.ProjectSystem.Tokenizer;
 
-namespace Microsoft.VisualStudio.Testing
+namespace Microsoft.VisualStudio.ProjectSystem
 {
     // Parses a string into a project tree
     //
