+++ src/Microsoft.VisualStudio.ProjectSystem.Managed.TestServices/ProjectSystem/ProjectTreeParser/Tokenizer.IdentifierParseOptions.cs
@@ -1,4 +1,4 @@
-﻿namespace Microsoft.VisualStudio.Testing
+﻿namespace Microsoft.VisualStudio.ProjectSystem
 {
     internal partial class Tokenizer
     {
