+++ src/Microsoft.VisualStudio.ProjectSystem.VisualBasic.UnitTests/ProjectSystem/MyProjectFolderProjectTreeModifierTests.cs
@@ -3,7 +3,6 @@
 using System;
 using Microsoft.VisualStudio.ProjectSystem.Imaging;
 using Microsoft.VisualStudio.ProjectSystem.Properties;
-using Microsoft.VisualStudio.Testing;
 using Xunit;
 
 namespace Microsoft.VisualStudio.ProjectSystem
